+++ src/index.css
@@ -12,13 +12,13 @@ code {
     monospace;
 }
 
-.container {
+div .container {
   display: flex;
   justify-content: center;
   align-items: center;
-  margin-top: 2rem;
+  margin: 1.5rem;
 }
 
 button{
-  margin: 1rem;
+  margin: 0.5rem;
 }
